requirements.txt
Flask
Flask-Caching
Flask-Cors
Flask-Limiter
Flask-SocketIO
flask-talisman
gunicorn
eventlet
mysql-connector-python
redis
bcrypt
PyJWT
python-dotenv
Pillow
email-validator
Werkzeug
python-socketio
python-engineio
cachelib
limits
blinker
itsdangerous
Jinja2
MarkupSafe
click
cryptography
requests
urllib3
certifi
charset-normalizer
idna
bleach